devided into each server
